Remove embeded mixin from the client/server itself.
This commit stops Mixin from being embedded into the client/server jar and instead embeds it into the loader. This allows for us to get rid of those nasty Mixin host service exceptions and it simply does not belong in the client as the client simply just uses Mixin like a normally mod would but does not bootstrap it etc.